fix: there was no operator= which caused trouble on pointers fix: various narrowing warning suppressions

+++ main.cpp
@@ -135,7 +135,7 @@ int main()
                 c << stamp(gout, px, py, 40,40,0,0);
                 cv.push_back(c);
             }
-            for (int i=0;i<cv.size();i++) {
+            for (size_t i=0;i<cv.size();i++) {
                 gout << stamp(cv[i], rand()%X-40,rand()%Y-40);
             }
             gout << refresh;
@@ -154,7 +154,7 @@ int main()
 
     gout << move_to(0,0) << color(0,0,0) << box(X,Y) << color(255,255,255) << refresh;
     vector<canvas> cv(10);
-    for (int i=0;i<cv.size();i++) {
+    for (size_t i=0;i<cv.size();i++) {
     	cv[i].open(100,100);
     	for (int j=0;j<10;j++) {
     		cv[i] << move_to(i*8,i*8) << line_to(100-j*7, j*7);
@@ -165,7 +165,7 @@ int main()
     while (gin>>ev && ev.keycode != key_escape) {
     	if (ev.type == ev_timer) {
 	    	gout << move_to(0,0) << color(0,0,0) << box(X,Y) << color(255,255,255);
-	    	for (int i=0; i<cv.size(); i++) {
+	    	for (size_t i=0; i<cv.size(); i++) {
 	    		gout << stamp(cv[i],X/2+sin(ev.time/1000.0+i)*X/3,Y/2+sin(ev.time/1200.0+i)*Y/3);
 	    	}
 	    	gout << refresh;
